archive/simple-mode-client/requirements.txt
livekit-api>=0.8.0
